# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "园智汇 · AI 产业园区智慧运营管理平台",
  "description": "SmartPark AI Operations Platform\n\n面向产业园区集团的多租户智慧运营平台，包含数据驾驶舱、空间资产、企业服务、招商运营、合同财务、项目管理、物业运维、安全管理、审批中心、AI Agent（1 个主 Agent + 11 个子 Agent）、数据中心、报表中心、日志审计与系统管理。\n\n**数据说明**：演示环境，全部数据为演示数据（is_demo=True），由数据库实时聚合计算。",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/dashboard/summary": {
   "get": {
    "tags": [
     "经营驾驶舱"
    ],
    "summary": "Summary",
    "operationId": "summary_api_dashboard_summary_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "park_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Park Id"
      }
     },
     {
      "name": "building_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Building Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Summary Api Dashboard Summary Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/dashboard/alerts": {
   "get": {
    "tags": [
     "经营驾驶舱"
    ],
    "summary": "Alerts",
    "description": "今日洞察与预警（来自通知表，按严重度排序）。",
    "operationId": "alerts_api_dashboard_alerts_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "park_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Park Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 20,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Alerts Api Dashboard Alerts Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/dashboard/quick-stats": {
   "get": {
    "tags": [
     "经营驾驶舱"
    ],
    "summary": "Quick Stats",
    "description": "顶部快捷统计（轻量，供全局框架使用）。",
    "operationId": "quick_stats_api_dashboard_quick_stats_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Quick Stats Api Dashboard Quick Stats Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/api/auth/login"
     }
    }
   }
  }
 }
}